change from rbdX to pool/image for the configuration dict, and enable sorts of the summary data
The key used for the configuration was changed from the rbd device (rbdX) to the pool/image name. This ensures that if the rbd image is assigned to a different rbdX device across nodes, the data will still aggregate and summarise correctly. In addition, invocation now accepts a sort key (-s) and a boolean for reverse sort of the data presented to the admin. By default the data is presented in pool/image ascending sequence but this can now be changed to rbd name, iops await etc