config.lio : correct client count bug and provide client alias in the UI output
It was possible for the client count to include a null entry, so this code was
rewritten to get the client count directly from rtslib.

Also, when showing the client name the name shown is the dns alias of the
client if it's connected, or the last qualifier of the iqn if not. A name of '- multi -'
is shown if a lun is connected to multiple clients