set the default pcp collector based on the lun types defined in LIO
The device_map dict now contains a lun_type attribute (user or block), so
by looking at the dict to see if user LUNs are present we can automatically
default to the correct PCP collector. A config needs to have just a single
user LUN to cause the LIO pcp collector to be used.